refactor: use barrel file instead of specific file
