+++ functions/src/modules/authenticators/infrastructure/repositories/access.repository.spec.ts
@@ -2,7 +2,7 @@ import "reflect-metadata";
 import {FirestoreAccessRepository} from "./access.repository";
 import {Access} from "../../domain/entities/access.entity";
 import {container} from "tsyringe";
-import {AuthenticatorType} from "../../domain/entities/authenticator-type.enum";
+import {AuthenticatorType} from "../../domain/entities";
 
 describe("FirestoreAccessRepository", () => {
   let repo: FirestoreAccessRepository;
